Moved editor into sugar3.graphics
